Rename JSON to schema
In order to prevent conflicts with the real JSON library
